requirements.txt
pyspark
cassandra-driver